// lib/contentful/__generated/sdk.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: { input: string; output: string };
  String: { input: string; output: string };
  Boolean: { input: boolean; output: boolean };
  Int: { input: number; output: number };
  Float: { input: number; output: number };
  DateTime: { input: any; output: any };
  Dimension: { input: any; output: any };
  HexColor: { input: any; output: any };
  JSON: { input: any; output: any };
  Quality: { input: any; output: any };
};

/** Represents a binary file in a space. An asset can be any file type. */
export type Asset = {
  __typename?: 'Asset';
  contentType?: Maybe<Scalars['String']['output']>;
  contentfulMetadata: ContentfulMetadata;
  description?: Maybe<Scalars['String']['output']>;
  fileName?: Maybe<Scalars['String']['output']>;
  height?: Maybe<Scalars['Int']['output']>;
  linkedFrom?: Maybe<AssetLinkingCollections>;
  size?: Maybe<Scalars['Int']['output']>;
  sys: Sys;
  title?: Maybe<Scalars['String']['output']>;
  url?: Maybe<Scalars['String']['output']>;
  width?: Maybe<Scalars['Int']['output']>;
};
// ...
export type AssetCollection = {
  __typename?: 'AssetCollection';
  items: Maybe<Asset>[];
  limit: Scalars['Int']['output'];
  skip: Scalars['Int']['output'];
  total: Scalars['Int']['output'];
};
// ...
export type AssetLinkingCollections = {
  __typename?: 'AssetLinkingCollections';
  componentAuthorCollection?: Maybe<ComponentAuthorCollection>;
  componentRichImageCollection?: Maybe<ComponentRichImageCollection>;
  componentSeoCollection?: Maybe<ComponentSeoCollection>;
  entryCollection?: Maybe<EntryCollection>;
  pageBlogPostCollection?: Maybe<PageBlogPostCollection>;
};
// ...
export type ComponentAuthor = Entry & {
  __typename?: 'ComponentAuthor';
  avatar?: Maybe<Asset>;
  contentfulMetadata: ContentfulMetadata;
  internalName?: Maybe<Scalars['String']['output']>;
  linkedFrom?: Maybe<ComponentAuthorLinkingCollections>;
  name?: Maybe<Scalars['String']['output']>;
  sys: Sys;
};
// ...
export type ComponentAuthorLinkingCollections = {
  __typename?: 'ComponentAuthorLinkingCollections';
  entryCollection?: Maybe<EntryCollection>;
  pageBlogPostCollection?: Maybe<PageBlogPostCollection>;
};
// ...
export type ComponentRichImage = Entry & {
  __typename?: 'ComponentRichImage';
  caption?: Maybe<Scalars['String']['output']>;
  contentfulMetadata: ContentfulMetadata;
  fullWidth?: Maybe<Scalars['Boolean']['output']>;
  image?: Maybe<Asset>;
  internalName?: Maybe<Scalars['String']['output']>;
  linkedFrom?: Maybe<ComponentRichImageLinkingCollections>;
  sys: Sys;
};
// ...
export type ComponentRichImageLinkingCollections = {
  __typename?: 'ComponentRichImageLinkingCollections';
  entryCollection?: Maybe<EntryCollection>;
};
// ...
export type ComponentSeo = Entry & {
  __typename?: 'ComponentSeo';
  canonicalUrl?: Maybe<Scalars['String']['output']>;
  contentfulMetadata: ContentfulMetadata;
  internalName?: Maybe<Scalars['String']['output']>;
  linkedFrom?: Maybe<ComponentSeoLinkingCollections>;
  nofollow?: Maybe<Scalars['Boolean']['output']>;
  noindex?: Maybe<Scalars['Boolean']['output']>;
  pageDescription?: Maybe<Scalars['String']['output']>;
  pageTitle?: Maybe<Scalars['String']['output']>;
  shareImagesCollection?: Maybe<AssetCollection>;
  sys: Sys;
};
// ...
export type ComponentSeoLinkingCollections = {
  __typename?: 'ComponentSeoLinkingCollections';
  entryCollection?: Maybe<EntryCollection>;
  pageBlogPostCollection?: Maybe<PageBlogPostCollection>;
  pageLandingCollection?: Maybe<PageLandingCollection>;
};
// ...
export type ContentfulMetadata = {
  __typename?: 'ContentfulMetadata';
  tags: Maybe<ContentfulTag>[];
};
// ...
export type ContentfulTag = {
  __typename?: 'ContentfulTag';
  id?: Maybe<Scalars['String']['output']>;
  name?: Maybe<Scalars['String']['output']>;
};

export type Entry = {
  contentfulMetadata: ContentfulMetadata;
  sys: Sys;
};

export type EntryCollection = {
  __typename?: 'EntryCollection';
  items: Maybe<Entry>[];
  limit: Scalars['Int']['output'];
  skip: Scalars['Int']['output'];
  total: Scalars['Int']['output'];
};
// ...
export type PageBlogPost = Entry & {
  __typename?: 'PageBlogPost';
  author?: Maybe<ComponentAuthor>;
  content?: Maybe<PageBlogPostContent>;
  contentfulMetadata: ContentfulMetadata;
  featuredImage?: Maybe<Asset>;
  internalName?: Maybe<Scalars['String']['output']>;
  linkedFrom?: Maybe<PageBlogPostLinkingCollections>;
  publishedDate?: Maybe<Scalars['DateTime']['output']>;
  relatedBlogPostsCollection?: Maybe<PageBlogPostRelatedBlogPostsCollection>;
  seoFields?: Maybe<ComponentSeo>;
  shortDescription?: Maybe<Scalars['String']['output']>;
  slug?: Maybe<Scalars['String']['output']>;
  sys: Sys;
  title?: Maybe<Scalars['String']['output']>;
};
// ...
export type PageBlogPostContent = {
  __typename?: 'PageBlogPostContent';
  json: Scalars['JSON']['output'];
  links: PageBlogPostContentLinks;
};

export type PageBlogPostContentAssets = {
  __typename?: 'PageBlogPostContentAssets';
  block: Maybe<Asset>[];
  hyperlink: Maybe<Asset>[];
};

export type PageBlogPostContentEntries = {
  __typename?: 'PageBlogPostContentEntries';
  block: Maybe<Entry>[];
  hyperlink: Maybe<Entry>[];
  inline: Maybe<Entry>[];
};

export type PageBlogPostContentLinks = {
  __typename?: 'PageBlogPostContentLinks';
  assets: PageBlogPostContentAssets;
  entries: PageBlogPostContentEntries;
  resources: PageBlogPostContentResources;
};

export type PageBlogPostContentResources = {
  __typename?: 'PageBlogPostContentResources';
  block: ResourceLink[];
  hyperlink: ResourceLink[];
  inline: ResourceLink[];
};
// ...
export type PageBlogPostLinkingCollections = {
  __typename?: 'PageBlogPostLinkingCollections';
  entryCollection?: Maybe<EntryCollection>;
  pageBlogPostCollection?: Maybe<PageBlogPostCollection>;
  pageLandingCollection?: Maybe<PageLandingCollection>;
};
// ...
export type PageBlogPostRelatedBlogPostsCollection = {
  __typename?: 'PageBlogPostRelatedBlogPostsCollection';
  items: Maybe<PageBlogPost>[];
  limit: Scalars['Int']['output'];
  skip: Scalars['Int']['output'];
  total: Scalars['Int']['output'];
};
// ...
export type PageLanding = Entry & {
  __typename?: 'PageLanding';
  contentfulMetadata: ContentfulMetadata;
  featuredBlogPost?: Maybe<PageBlogPost>;
  internalName?: Maybe<Scalars['String']['output']>;
  linkedFrom?: Maybe<PageLandingLinkingCollections>;
  seoFields?: Maybe<ComponentSeo>;
  sys: Sys;
};
// ...
export type PageLandingLinkingCollections = {
  __typename?: 'PageLandingLinkingCollections';
  entryCollection?: Maybe<EntryCollection>;
};
// ...
export type Sys = {
  __typename?: 'Sys';
  environmentId: Scalars['String']['output'];
  firstPublishedAt?: Maybe<Scalars['DateTime']['output']>;
  id: Scalars['String']['output'];
  publishedAt?: Maybe<Scalars['DateTime']['output']>;
  publishedVersion?: Maybe<Scalars['Int']['output']>;
  spaceId: Scalars['String']['output'];
};
// ...
export type AuthorFieldsFragment = {
  __typename: 'ComponentAuthor';
  name?: string | null;
  sys: { __typename?: 'Sys'; id: string };
  avatar?: ({ __typename?: 'Asset' } & ImageFieldsFragment) | null;
};

export type ImageFieldsFragment = {
  __typename: 'Asset';
  title?: string | null;
  description?: string | null;
  width?: number | null;
  height?: number | null;
  url?: string | null;
  contentType?: string | null;
  sys: { __typename?: 'Sys'; id: string };
};

export type ReferencePageBlogPostFieldsFragment = {
  __typename: 'PageBlogPost';
  slug?: string | null;
  publishedDate?: any | null;
  title?: string | null;
  shortDescription?: string | null;
  sys: { __typename?: 'Sys'; id: string; spaceId: string };
  author?: ({ __typename?: 'ComponentAuthor' } & AuthorFieldsFragment) | null;
  featuredImage?: ({ __typename?: 'Asset' } & ImageFieldsFragment) | null;
};

export type PageBlogPostFieldsFragment = {
  __typename: 'PageBlogPost';
  internalName?: string | null;
  slug?: string | null;
  publishedDate?: any | null;
  title?: string | null;
  shortDescription?: string | null;
  sys: { __typename?: 'Sys'; id: string; spaceId: string };
  seoFields?: ({ __typename?: 'ComponentSeo' } & SeoFieldsFragment) | null;
  author?: ({ __typename?: 'ComponentAuthor' } & AuthorFieldsFragment) | null;
  featuredImage?: ({ __typename?: 'Asset' } & ImageFieldsFragment) | null;
  content?: {
    __typename?: 'PageBlogPostContent';
    json: any;
    links: {
      __typename?: 'PageBlogPostContentLinks';
      entries: {
        __typename?: 'PageBlogPostContentEntries';
        block: Array<
          | { __typename?: 'ComponentAuthor' }
          | ({ __typename?: 'ComponentRichImage' } & RichImageFieldsFragment)
          | { __typename?: 'ComponentSeo' }
          | { __typename?: 'PageBlogPost' }
          | { __typename?: 'PageLanding' }
          | null
        >;
      };
    };
  } | null;
  relatedBlogPostsCollection?: {
    __typename?: 'PageBlogPostRelatedBlogPostsCollection';
    items: Array<
      | ({ __typename?: 'PageBlogPost' } & ReferencePageBlogPostFieldsFragment)
      | null
    >;
  } | null;
};
// ...
export type RichImageFieldsFragment = {
  __typename: 'ComponentRichImage';
  internalName?: string | null;
  caption?: string | null;
  fullWidth?: boolean | null;
  sys: { __typename?: 'Sys'; id: string };
  image?: ({ __typename?: 'Asset' } & ImageFieldsFragment) | null;
};

export type SeoFieldsFragment = {
  __typename: 'ComponentSeo';
  pageTitle?: string | null;
  pageDescription?: string | null;
  canonicalUrl?: string | null;
  nofollow?: boolean | null;
  noindex?: boolean | null;
  shareImagesCollection?: {
    __typename?: 'AssetCollection';
    items: Array<({ __typename?: 'Asset' } & ImageFieldsFragment) | null>;
  } | null;
};